Try a new strategy for safe GcObject trait
The old design was to have GcObject be unsafe, but Gc pointers would be truly
zero-cost, they would just be a machine pointer.  The idea was that using this,
you could make safe containers that exposed rooted pointers on the outside and
would be safe externally.  The problem with this in the presence of separate gc
contexts is that you would have to check at the boundaries of each container
that accepted inserting values that the pointers you recieved were actually from
the same gc context, otherwise this would result in unsafety.  This is a problem
due to both space and time overhead, because you need to keep context pointers
in every GcBox, and also perform comparison checks.

Given the requirements to do this checking on the inside of unsafe containers, I
came up with a strategy that is probably not much worse if at all, but allows
for a completely safe interface to the gc.  Gc pointers now do reference
counting as well as Rc pointers, and the gc context prevents freeing the Gc box
until the last Gc pointer is destructed (it will still destruct the contained
value upon sweep, as normal).  This allows for "dangling" Gc pointers to safely
panic, so suddenly the GcObject trait is also safe.  Since this is safe, it
becomes safe to ignore gc context checks and also expose plain Gc pointers in
interfaces.

Where I've ended up is very close to just using Rc with cycle detection, sort
of.  I think doing it this way eliminates the inline checks to trigger
destruction on both Gc and Rgc, so maybe that's something, but I'm not sure how
much different that is in practice.  Even if the overhead is low, this is sort
of disheartening because Rc is not *that* great, and there is certainly more
bookkeeping than PUC-Rio Lua.  This is not very far even from something built
using Rc / Weak and entirely safe code, which is maybe interesting academically
that you can build a safe Gc, but I just don't know if that's going to be fast
enough.

Either the overhead here is going to be small and I'm worrying over nothing, or
the overhead is not small at all and in order to be competitive the unsafety of
the Gc must infect the entire interpreter.