I *believe* this table growth logic matches vanilla Lua's closely enough
The only caveat is that I think vanilla Lua will actually shrink both the array
and map parts, but it only shrinks the array part when the map part exceeds its
capacity, and shrinking the map part is done every time because it simply
removes map nodes rather than having to resize a contiguous array like rust's
HashMap.  This implementation simply never shrinks the array or map parts, but
it would be reasonable to expose reserving / shrink to fit on the table somehow.