Redesign of register allocation... again
PUC-Rio Lua's compiler is carefully delicately designed so that registers which
are allocated are always allocated at the end, and registers which are freed are
always freed in reverse order of allocation.  This is delicate and easy to get
wrong.  However, registers cannot only be freely allocated, because many things
assume that register allocation occurs at the top of the stack, and assume all
upper values are unused.

This creates a register allocator with two concepts: "allocate" and "free", and
"push" and "pop", which can be mixed, and allows for freeing registers in any
order.