Add function call expression descriptor to compiler
The logic for why this is necessary is somewhat subtle.  PUC-Rio Lua handles function calls during
code generation by generating the function call and assuming a return count of 1, then later on
during parsing / code generation it updates the return count with the proper return count which it
has at that point discovered.  The code generator knows this is safe simply because it is structured
so that nothing that uses additional registers will happen between the initial opcode generation and
the later update.  This is rather scary and implicit, especially with long chains or function calls
with some in the last position (this enables vararg handling).  It seems to work fine, but it is
at least somewhat non-obvious.

This is an alternative approach which simply delays codegen of function calls until enough
information is gathered so that all of the opcodes can be generated without further update.  This
requires, for example, parsing full function call expressions to know whether an inner function call
is in the final position and should return MULTRET.  Without doing things this way, the compiler
would either have to:

- Generate a 1 return function call then update it later, like PUC-Rio Lua, which would rely on the
  inherent structure of the compiler to not do incompatible things in between generating the
  function call opcode and updating it afterwards, and potentially be very confusing with the
  register allocator.
- Pass through expression handling the expected return count, which would then require a very
  awkward state where there exists an `ExprDescriptor` that marks that a variable return had just
  occurred, and would still require that it was immediately applied to another function call to
  avoid code generation bugs (which is also true in PUC-Rio Lua).

Having `ExprDescriptor::FunctionCall` and waiting until the result is actually used to generate the
call opcode is simpler all around and much more obviously correct, it just requires slightly more
allocation.  It is not a *current* goal of luster to match PUC-Rio's extremely low memory usage
during compilation, but even if it was there are lots of other changes which would be much, much
more fruitful.