I remember now, the idea was to clear the peek_buffer on error
But definitely NOT on simple eof.  This is not so important, as the idea was that if the lexer
entered an "error state" it would act as though it had encountered eof, but there are several other
errors that should also cause this state as well then, and it should be handled outside of peek.