Preparation for proper outer api and better thread handling:
Make upvalues correct in the presence of multiple threads by having them include their parent thread
as well as the stack index.

In order not to get multiple mutable borrow errors, the stack checks whether the target stack is the
current stack before borrowing.

This design implies that no more than one thread can ever be mutably borrowed at once, and
therefore that threads cannot directly call methods on other threads.  This is fine, however,
because this limtation was already a design constraint, and this fits within the current tentative
design.

Soon, running Lua bytecode on a Thread will have to have a sort of continuation-return-style.
Trying to execute a given number of bytecode instructions is not a good strategy, what instead has
to happen is that Thread::run needs to cooperatively return when it needs certain actions to take
place.  It should return with a `CollectGarbage` code to signal that a gc is needed, it should
return with a `Callback(...)` to signal that a callback is needed, and it should return with
`CoroutineResume(...)` or something to signal that a different thread should be entered.  Something
outside the Thread, and at least partially outside the GcArena entirely, needs to be running the
thread and handling the operations returned.  This is the only way that garbage collection can be
performed, since it cannot be performed inside the GcArena::mutate call, and without structuring
callbacks in this way, callbacks would always have to block garbage collection for their entire
duration.  The details are still a bit fuzzy here, but at the very least it doesn't seem any harder
to move to a different coroutine via return than it does callbacks or garbage collection.