Redesign expression handling in the compiler to fix expr push bugs
Collapsed all of the expression handling functions into "expr_discharge", which
optionally produces a result in a register.  This is necessary in order to
correctly handle discharging into a result on the top of the stack, since we
want to be on the top of the stack *after* freeing all of the used registers in
the expression.