Implement gotos with labels (untested and known incorrect)
This is wrong, because I misunderstood how Lua labels work.  Labels *are*
allowed to have duplicates within a function if the labels are not within the
same block level, and inner labels are not visible to outer gotos.  There is no
reason for the JumpBlock error, as it should always be equivalent to not finding
a matching label.  Fixing, and will test after this.