Implement while loops and correct if statements
I believe this implements if statements and while loops semantically
correctly (ish?), but it generates horrible code in comparison to PUC-Rio Lua,
which understands at least one level of `if` inside `while` nesting and
eliminates redundant jumps, and has special cases for handling if clauses with
break and goto.  I need to figure out how to generate better code here.