Fix closing upvalues with while loops by generating meh bytecode
Several things will need to change very soon, and I'm going to write my thoughts
here just to convince myself I understand it:

First, this fix generates bad bytecode because it does not coalesce the upvalue
closing no-op jump with the jump following it, and it doesn't do this because it
is extremely *hard* the way the code is structured to do so.

In order to have the end jump automatically close upvalues itself, it needs to
be inside the given block (it would be inside the block jumping out, and so the
jump would automatically close that blocks upvalues).  Right now, the end jump
takes place *right after* the block closes, because it has to: it has to follow
the block's "trailing labels" that have special behavior in Lua.  The trailing
labels *can't* go inside the block because they require that the locals in the
block be closed in order not to generate errors from jumping into a new local's
scope.  This is a catch 22: the end jump needs to go before the block closes but
after the trailing labels, and the trailing labels must go after the block
closes.

In order to fix this, the way jumps and jump targets work would need special
cases for ignoring local variable scope, or some other similar special case
solution.  That is not *so* bad, but this entire solution is in already in the
wrong direction and can't be made to match PUC-Rio Lua anyway!  The simplest way
this is wrong is that probably 'repeat / until' loops can't be made to work with
this pattern, because the until clause in a 'repeat / until' is actually in
scope of the block.  The block handling code is probably going to have to be
split up into block_enter / statlist / block_exit like PUC-Rio Lua does.

There is a more serious way the current approach is wrong though... take the
following code:

```
do
    goto next1
    ::next1::
    goto next2
    ::next2::
    goto next3
    ::next3::
end
```

PUC-Rio Lua compiles this into the following bytecode:

```
        1       [1]     JMP             0 2     ; to 4
        2       [3]     JMP             0 1     ; to 4
        3       [5]     JMP             0 0     ; to 4
        4       [8]     RETURN          0 1
```

The reason for this is that PUC-Rio Lua keeps unresolved jumps in lists, so that
when one jump goes immediately to another they become chained, and when they are
resolved they are all resolved to jump directly to the final location.  With
this technique, there is no real need to fold jumps together as I was
attempting, because redundant jumps would be automatically skipped.

The next step is probably to implement jump chaining, and then do while / until
loops because it is the most complicated scoping case