Stop using "local count" to mean "stack top"
Theoretically this was a bug before, but it would never be hit in practice
because the "local count" was always equal to the stack top, because there
shouldn't have been any allocation holes.

However, this wasn't enforced and it was confusing and is actually *incorrect*
in the presence of allocated registers that aren't assigned local
variables (like in for loops).

Now, all local variables are pushed instead of "allocated" for clarity (there
should be no allocation holes, maybe there should be an assert for this?).
Also, all variables referencing the local count now instead reference the stack
top instead, which is required for for loops to work.