WIP refactoring, make ExprDescriptor never actually execute code
The goal here is to make ExprDescriptor an intermediate representation of an
expression rather than a result of partially evaluating an expression.  This
way, execution order can be controlled sanely, and much more optimal bytecode
can be generated.

This is a partial step where ExprDescriptor contains the IR as well as being
able to hold an evaluated expression in a register.  The next step is to remove
that.