Huge refactor to move VM code out of Thread
Improves Thread abstraction, but introduces a lot of new, repeating code.  This
is just a first step, hopefully Thread can be further simplified again.

Fixes at least a few bugs, but probably introduces different bugs.

As a drive-by, added `SequenceExt::flatten` to better deal with things that
produce sequences with the chance of error.