On the balance, it's actually less complicated without `bin_op` now
This is sort of a judgement call based on the idea that certain operations
involving bin_op are about to get more complicated.

Observe this behavior in PUC-Rio Lua:

```
> 9223372036854774785 < 9223372036854774786.0
false
> 9223372036854774785 < 9223372036854774784.0
false
> 9223372036854774784 < 9223372036854774784.0
false
> 9223372036854774784 < 9223372036854774785.0
false
> 9223372036854774784 < 9223372036854774785
true
> 9223372036854774784.0 < 9223372036854774785
true
> 9223372036854774786.0 < 9223372036854774785
true
> 9223372036854774786.0 < 9223372036854774785.0
false
```

I believe `<` is not as simple as casting both the left and right to a number,
and the same is true I believe with equals, greater than, etc (I also believe
the current implementation of Value::partial_eq is wrong).