Couple of changes:
Remove Display impl for Gc, move Debug impl for Gc / GcCell to ones that *don't*
display their contents (only the pointer value).  The reason for this is that
the whole point of Gc is that it is normal and allowable to create cycles, so
this is kind of a bad idea.

Combines parser / compiler, use clap for compiler arguments (very basic right
now).  Also moves FunctionProto printing code into compiler.

Related to the above Gc change, fix Debug impl for String to be at least
marginally useful, and print the utf8 string if it is in fact utf8.