Better error handling through anyhow and dynamic errors
Because this crate is being designed primarily to integrate into a
larger Rust program, we need a way to support dynamic Rust errors, and
`anyhow` is the best and easiest way to do this.

As a side effect, dramatically simplifies error handling as well as
making it possible to transport Rust errors transparently through Lua.