Add a new `Execution` type passed to callbacks
Contains the old `Fuel` parameter as well as the current thread and
whether or not the current thread is the "main" thread.

This is also important so that in the near future this can be used to
get further execution context like the current file / line number and
also stack traces.