More bugfixes for weak thread finalization
If a thread was held in an upvalue, then it was possible for the runtime
to think that this thread was dead when it was not and reset it.

To prevent this, the runtime has been changed to mark the *targets* of
all live upvalues as live during finalization, even though they are held
through a weak thread ptr.

This requires two-phase finalization, one to make sure all live upvalues
are actually marked as live, *then* one to reset every dead thread.
Multi-phase finalization has always been on the roadmap so this is not
really a problem.