Allow `Sequence` impls to call/yield/resume at an offset other than 0
Previously, when a `Sequence` would call a function, yield values, or
resume a thread, every value currently in the stack would be used as the
arguments. Now instead, a `bottom` may be specified that may be
non-zero, and the arguments for the chosen operation instead start at
`bottom`. Return values will also be placed at `bottom` rather than
replacing the entire stack, and in case of an error, the stack will
still contain values before `bottom` when `Sequence::error` is called.
Specifying a `bottom` value of 0 produces the previous behavior.

*Tail* call/yield/resume will always use the entire stack, and thus
those have been separated into their own `SequencePoll` variants.