Always do at least one batch in `Unpack` sequence
`Executor::step` should always make minimal progress, even when out of
fuel, so we need to break from the loop at the bottom, not the top.

Also gets rid of the outer loop in Unpack::poll for clarity with
breaking for out-of-fuel, instead doing manual state transitions at the
beginning.