Controversial move: remove the hard-coded limit on table.unpack
It is no longer strictly *required*, because no matter how big the value
is, control will return to Rust after some smaller allocation that is
proportional to the amount of fuel given.

It is very reasonable to want stack limits or allocation limits in
general, but I would rather leave this for something more universal.