Add a lot of rustdoc links and clarify comments some more
Removes a NOTE: in the doc comments for `Locals` that is not very
relevant because 1) literally calling `mem::forget` on a `Local` is
degenerate and silly to consider, and 2) you actually *can't* create
`Local` cycles because `Local` cannot be placed inside Lua values (due
to the 'seq branding). You can't mess up with `Local` as much as you can
values stashed in the global registry.