Implement async sequences (#84)
* Implement async sequences

* don't use Rc when we don't have to

* simplify async sequences a small amount by getting rid of SeqContext

* add a lot of documentation to async sequences

* More documentation, some small re-org

Reorganization makes the diff from master smaller and makes a bit more
sense to me

* small comment change

* Be much more strict about the behavior of sequence futures

If the rules are not followed, rather than trying to guess at something
to do, just panic.

Also adds comments to better explain the situation and when panics may
occur.

* more comment changes

* spelling fixes

* spelling fixes and rewords

* Add a lot of rustdoc links and clarify comments some more

Removes a NOTE: in the doc comments for `Locals` that is not very
relevant because 1) literally calling `mem::forget` on a `Local` is
degenerate and silly to consider, and 2) you actually *can't* create
`Local` cycles because `Local` cannot be placed inside Lua values (due
to the 'seq branding). You can't mess up with `Local` as much as you can
values stashed in the global registry.