Simplify the interface and implementation of `AsyncSequence` by making one change
Instead of keeping the `create` function around and calling it on first
poll, simply calls it immediately.

We no longer need `AsyncSequence::new_sequence_with`, because the
callback in `AsyncSequence::new_sequence` is not required to be 'static.

This also allows us to get rid of the `SeqFut` enum entirely, as there
is no pending create variant to worry about.

This also makes the implementation easier to explain and understand in
general, as the new_sequence callback is now ONLY a callback so that we
can produce a generative 'seq lifetime, and we can explain things this
way.

The only change is that we need to document that the `create` callback
must not call methods on the provided `SequenceState` outside of the
returned future, but this actually requires no extra code to accomplish,
as we simply don't call it within the context of a `with_shared`, so
`visit_shared` will naturally panic.