Avoid double boxing in async sequences by changing the API
Instead of providing a create function that returns a future, instead
give the create method a "builder" that can create `BoxFuture` directly.
Since the create function has no return values, we don't have to worry
about the impossibility of expressing that the return type is a future
AND it is bound to the HRTB 'seq.

Avoidiing double boxing also requires changing the `Sequence` trait to
accept `self: Pin<&mut Self>`, which requires changing the signature of
`Sequence` impls *everywhere* and this is somewhat invasive, but it is
not actually a difficult thing to deal with as a `Sequence` implementer.
It is also trivial to make work internally, as every sequence is already
stored as a `BoxSequence` anyway.