Combine StashedXXX values and LocalXXX values into a single type
All StashedXXX values are now branded by an invariant 'ctx lifetime,
values stashed in the registry have 'ctx set to 'static to make them
unrestricted.

This is a small usability setback, because values stashed in the
registry like `StashedExecutor` need to be changed to
`StashedExecutor<'static>`. We could provide a separate set of typedefs
that are always 'static if this is too annoying.

The benefit is that there is not a split between Local and Stashed
values, meaning that Local values no longer have to be opaque wrappers
or differently shaped than their stashed counterparts.