More changes to the async sequences API
Simplify everything by ditching the entire 'seq lifetime idea.

It provided a benefit, but the benefit was miniscule in comparison to
the complexity cost. The main benefit was that it prevented misusing
stashed values when stashed inside the future driving an async sequence.
HOWEVER, several things already prevent this misuse:

One, the futures must be 'static (they were 'seq, but this is the same
idea), so the only way to tunnel a Stashed handle out of the future is
by using something like a channel or an Rc, and this is pretty hard to
do *on accident*.

Two, `DynamicRoot` and thus all of the stashed handle types do not
implement `Collect`, and this *already* prevents storing any of these
roots inside the GC context without at least some extra work. You can
make a new type which implements Collect and stores a root and
`#[collect(require_static)]` that field, but this is already an extra
speed bump. You can always capture one of these handles in a callback or
something too, but the handles are completely unusable without the
`AsyncSequence` to fetch them, so again this is unlikely to be done on
accident.

The biggest issue though is that the user already has to be careful with
what they do because stashing and fetching values from the *registry* is
utterly painless *and wrong*. Since it's so easy to make the same kinds
of mistakes using the registry, and pretty tough to accidentally make
the mistakes that the 'seq lifetime prevented without using the
registry, the overall benefit ends up being small.

It is just as possible now as it was before to build an async future
correctly, and without the possibility of uncollectable GC cyles... only
a few guard rails have been removed. The result requires less unsafe
code to implement and less complexity for the user.