Make `FreezeGuard` more publicly useful
`FreezeGuard` is now effectively a single-element `FrozenScope` and is a
bit easier to understand.

As a side effect, we can now provide a way to get the value back *out*
of a `FreezeGuard`, since it only contains a single value.

This also includes some more internal changes to let
`FrozenScope::scope` and the new `FreezeGuard::scope` be called more
than once for the *same* object. Rather than requiring that
`FreezeGuard` be *dropped*, we instead add a `ScopeGuard::unset` call
which should unset the value *in a way that allows it to be set again*.

This would all be so much simpler with linear types.