Avoid monomorphizing entire compiler for each T: Read in load
The parser also has the same issue, but it isn't easily fixable
without tradeoffs.  However, duplicating the entire parser for
each input type and each crate is still a major problem.
(Currently there is a minimum of 2 instantiations of the entire
parser, once from the Lua stdlib and once for the user program,
assuming it only passes byte slices to Closure::load.)