feat: basic bench harness
Adds a basic Criterion-based benchmark harness which loads scripts
from the `benches/scripts` directory and times their execution as if
they were regular Rust functions.
The Criterion settings were tuned for fast runs, at the expense of
greater run-to-run variability. It is highly recommended that you
pin the benchmark to a specific CPU core to reduce the noise.